docs: update README and reorganize project structure
