+++ .env.example
@@ -1,7 +1,7 @@
-DATABASE_URL=
+DATABASE_URL=your_database_url
 
-JWT_SECRET=
+JWT_SECRET=your_secret
 
 PORT=3000
 
-SALT_PASSWORD=
+SALT_PASSWORD=10
