+++ src/modules/auth/auth.module.ts
@@ -5,7 +5,7 @@ import { AuthService } from '@auth/auth.service';
 import { JwtModule } from '@nestjs/jwt';
 import jwt from '@config/jwt';
 import { PrismaModule } from '@shared/prisma/prisma.module';
-import { JwtStrategy } from '@auth/guard/jwt.strategy';
+import { JwtStrategy } from '@auth/strategies/jwt.strategy';
 import { PassportModule } from '@nestjs/passport';
 
 @Module({
