+++ src/modules/user/dto/index.ts
@@ -1,5 +1,5 @@
-export * from '@user/dto/input/create-user.input.dto';
-export * from '@user/dto/input/update-user.input.dto';
-export * from '@user/dto/input/update-password.input.dto';
-export * from '@user/dto/input/find-user.input.dto';
-export * from '@user/dto/output/response-user.output.dto';
+export * from '@user/dto/inputs/create-user.input.dto';
+export * from '@user/dto/inputs/update-user.input.dto';
+export * from '@user/dto/inputs/update-password.input.dto';
+export * from '@user/dto/inputs/find-user.input.dto';
+export * from '@user/dto/outputs/response-user.output.dto';
